requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
attrs==25.1.0
bip-utils==2.9.3
bitarray==3.0.0
cbor2==5.6.5
cdp-sdk==0.16.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
coincurve==20.0.0
crcmod==1.7
cryptography==44.0.0
cytoolz==1.0.1
distro==1.9.0
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-core==0.3.33
langchain-text-splitters==0.3.5
langsmith==0.3.6
multidict==6.1.0
noise==1.2.2
numpy==1.26.4
opensimplex==0.4.5
openai==1.61.1
orjson==3.10.15
packaging==24.2
parsimonious==0.10.0
pillow==10.2.0
propcache==0.2.1
py-sr25519-bindings==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==9.0.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
zstandard==0.23.0
pytest==8.0.0
